fix compile err

+++ cmd/gk/command/dbus.go
@@ -74,7 +74,7 @@ func (this *Dbus) checkMyslave(zkzone *zk.ZkZone) {
 
 			// an orphan binlog stream: no dbus consuming it
 			lines = append(lines, fmt.Sprintf("-|%s|%s|%s|-|-|-|-|-", db,
-				v.File, gofmt.Comma(v.Offset)))
+				v.File, gofmt.Comma(int64(v.Offset))))
 			continue
 		}
 
@@ -94,7 +94,7 @@ func (this *Dbus) checkMyslave(zkzone *zk.ZkZone) {
 			if r == string(owner) {
 				ownerTime := zk.ZkTimestamp(ownerStat.Mtime).Time()
 				lines = append(lines, fmt.Sprintf("%s|%s|%s|%s|%s|%s|%s|%s|%s",
-					v.Name, db, v.File, gofmt.Comma(v.Offset), instance, string(ver), pid,
+					v.Name, db, v.File, gofmt.Comma(int64(v.Offset)), instance, string(ver), pid,
 					gofmt.PrettySince(uptime), gofmt.PrettySince(ownerTime)))
 			} else {
 				lines = append(lines, fmt.Sprintf("%s|%s| | |%s|%s|%s|%s| ",
